requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
httpx==0.25.2
passlib[argon2]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
google-auth==2.24.0
python-decouple==3.8
python-dateutil==2.8.2
